+++ pkg/services/object/put/distibuted_test.go
@@ -40,6 +40,7 @@ func (x testNetwork) IsLocalNodePublicKey(pk []byte) bool { return bytes.Equal(x
 
 func (x testNetwork) GetContainerNodes(cid.ID) (ContainerNodes, error) { panic("unimplemented") }
 func (x testNetwork) GetEpochBlock(uint64) (uint32, error)             { panic("unimplemented") }
+func (x testNetwork) GetEpochBlockByTime(uint32) (uint32, error)       { panic("unimplemented") }
 
 type testWorkerPool struct {
 	nCalls int